+++ R/pmcmc.R
@@ -22,30 +22,16 @@
 ##'   
 ##' @param n_mcmc number of mcmc mcmc iterations to perform
 ##' 
-##' @param pars_to_sample names of parameters to be sampled (currently beta_start, beta_end, beta_pl, and start_date,  gamma_triage, 
-##' gamma_hosp_R, gamma_hosp_D, gamma_ICU_R, gamma_ICU_D, and gamma_stepdown)
-##' 
-##' @param pars_init named list of initial inputs for parameters being sampled 
-##' 
-##' @param pars_min named list of lower reflecting boundaries for parameter proposals
-##' 
-##' @param pars_max named list of upper reflecting boundaries for parameter proposals
-##' 
+##' @param pars_to_sample \code{data.frame} detailing parameters to sample. Must contain columns
+##'   'names' (parameter names), 'init' (initial values), 'min' (minimum values), 'max' (maximum values),
+##'   'discrete (boolean indicating whether a discrete quantity)'
+##'                   
+##' @param pars_lprior functions to calculate log prior for each parameter. A named list for each parameter
+##'   listed in \code{pars_to_sample}. Each value must be a function which takes named parameter vector as 
+##'   input, returns a single numeric which is log of the prior probability.
+##'       
 ##' @param proposal_kernel named matrix of proposal covariance for parameters 
-##' 
-##' @param pars_discrete named list of logicals, indicating if proposed jump should be discrete
-##' 
-##' @param log_likelihood function to calculate log likelihood, must take named parameter vector as input, 
-##'                 allow passing of implicit arguments corresponding to the main function arguments. 
-##'                 Returns a named list, with entries:
-##'                   - $log_likelihood, a single numeric
-##'                   - $sample_state, a numeric vector corresponding to the state of a single particle, chosen at random, 
-##'                   at the final time point for which we have data.
-##'                   If NULL, calculated using the function calc_loglikelihood.
-##'                   
-##' @param log_prior function to calculate log prior, must take named parameter vector as input, returns a single numeric.
-##'                  If NULL, uses uninformative priors which do not affect the posterior
-##'                   
+##'
 ##' @param n_particles Number of particles
 ##' 
 ##' @param steps_per_day Number of steps per day
@@ -92,6 +78,7 @@
 ##'   
 ##' @export
 ##' @import coda 
+##' @import lubridate
 ##' @importFrom stats rnorm
 ##' @importFrom mvtnorm rmvnorm
 pmcmc <- function(data,
@@ -99,7 +86,8 @@ pmcmc <- function(data,
                   sircovid_model,
                   model_params,
                   pars_obs,
-                  pars_to_sample = c('beta_start',
+                  pars_to_sample = data.frame(
+                                    names=c('beta_start',
                                      'beta_end', 
                                      'beta_pl',
                                      'start_date',  
@@ -109,109 +97,147 @@ pmcmc <- function(data,
                                      'gamma_ICU_R', 
                                      'gamma_ICU_D', 
                                      'gamma_stepdown'),
-                  pars_init = list('beta_start'     = 0.14, 
-                                   'beta_end'       = 0.14*0.238,
-                                   'beta_pl'        = 0.14*0.238,
-                                   'start_date'     = as.Date("2020-02-07"),
-                                   'gamma_triage'   = 0.5099579,
-                                   'gamma_hosp_R'   = 0.1092046,
-                                   'gamma_hosp_D'   = 0.2911154,
-                                   'gamma_ICU_R'    = 0.3541429,
-                                   'gamma_ICU_D'    = 0.2913861,
-                                   'gamma_stepdown' = 0.452381),
-                  pars_min = list('beta_start'     = 0, 
-                                  'beta_end'       = 0,
-                                  'beta_pl'        = 0,
-                                  'start_date'     = 0,
-                                  'gamma_triage'   = 0,
-                                  'gamma_hosp_R'   = 0,
-                                  'gamma_hosp_D'   = 0,
-                                  'gamma_ICU_R'    = 0,
-                                  'gamma_ICU_D'    = 0,
-                                  'gamma_stepdown' = 0),
-                  pars_max = list('beta_start'     = 1, 
-                                  'beta_end'       = 1,
-                                  'beta_pl' =1,
-                                  'start_date'     = 1e6,
-                                  'gamma_triage'   = 1,
-                                  'gamma_hosp_R'   = 1,
-                                  'gamma_hosp_D'   = 1,
-                                  'gamma_ICU_R'    = 1,
-                                  'gamma_ICU_D'    = 1,
-                                  'gamma_stepdown' = 1
-                                  ),
+                                    init=c(0.14, 
+                                           0.14*0.238,
+                                           0.14*0.238,
+                                           lubridate::yday("2020-02-07"),
+                                           0.5099579,
+                                           0.1092046,
+                                           0.2911154,
+                                           0.3541429,
+                                           0.2913861,
+                                           0.452381),
+                                    min=c(0,
+                                          0,
+                                          0,
+                                          0,
+                                          0,
+                                          0,
+                                          0,
+                                          0,
+                                          0,
+                                          0),
+                                    max=c(1,
+                                          1,
+                                          1,
+                                          1e6,
+                                          1,
+                                          1,
+                                          1,
+                                          1,
+                                          1,
+                                          1),
+                                    discrete=c(FALSE,
+                                               FALSE,
+                                               FALSE,
+                                               TRUE,
+                                               FALSE,
+                                               FALSE,
+                                               FALSE,
+                                               FALSE,
+                                               FALSE,
+                                               FALSE),
+                                    stringsAsFactors = FALSE),
+                  pars_lprior = list('beta_start'     = function(pars) log(1e-10),
+                                     'beta_end'       = function(pars) log(1e-10),
+                                     'beta_pl'       = function(pars) log(1e-10),
+                                     'start_date'     = function(pars) log(1e-10),
+                                     'gamma_triage'   = function(pars) log(1e-10),
+                                     'gamma_hosp_R'   = function(pars) log(1e-10),
+                                     'gamma_hosp_D'   = function(pars) log(1e-10),
+                                     'gamma_ICU_R'    = function(pars) log(1e-10),
+                                     'gamma_ICU_D'    = function(pars) log(1e-10),
+                                     'gamma_stepdown' = function(pars) log(1e-10)),
                   proposal_kernel,
-                  pars_discrete = list('beta_start'     = FALSE,
-                                       'beta_end'       = FALSE,
-                                       'beta_pl'        = FALSE,
-                                       'start_date'     = TRUE,
-                                       'gamma_triage'   = FALSE,
-                                       'gamma_hosp_R'   = FALSE,
-                                       'gamma_hosp_D'   = FALSE,
-                                       'gamma_ICU_R'    = FALSE,
-                                       'gamma_ICU_D'    = FALSE,
-                                       'gamma_stepdown' = FALSE),
-                  log_likelihood = NULL,
-                  log_prior = NULL,
                   n_particles = 1e2,
                   steps_per_day = 4, 
                   output_proposals = FALSE,
                   n_chains = 1) {
+
+  if(length(output_proposals) != 1 || !is.logical(output_proposals)) {
+    stop("output_proposals must be either TRUE or FALSE")
+  }
+  
+  if (length(model_params$beta_y) > 1) {
+    stop("Set beta variation through generate_beta_func in sircovid_model, not model_params")
+  }
   
   #
   # Check pars_init input
   #
-  par_names <- c('beta_start', 'beta_end', 'beta_pl', 'start_date', 
-                 'gamma_triage', 'gamma_hosp_R', 'gamma_hosp_D', 
-                 'gamma_ICU_R',  'gamma_ICU_D', 'gamma_stepdown')
-  par_names <- par_names[par_names %in% pars_to_sample]
+  if (!all(c("names", "init", "min", "max", "discrete") %in% colnames(pars_to_sample))) {
+    stop("pars_to_samples must contain columns 'names', 'init', 'min', 'max' and 'discrete'")
+  }
+  par_names <- as.character(pars_to_sample$names)
   
   if (!all(c('beta_start', 'start_date') %in% par_names)) {
     stop("Turning off beta and start date sampling unsupported")
   }
-  
-  correct_format_vec <- function(pars) {
-      is.list(pars) & setequal(names(pars), 
-                             par_names)
+  if(!all(names(pars_lprior) %in% par_names)) {
+    stop("All sampled parameters must have a defined prior")
   }
   
-  correct_format_mat <- function(pars) {
+  correct_format_mat <- function(pars, names) {
        setequal(rownames(pars),
-                colnames(pars)) & 
+                colnames(pars)) && 
          setequal(rownames(pars),
-                   par_names)
+                   names) &&
+      nrow(pars) == length(names) &&
+      ncol(pars) == length(names)
   }
-  
-  if(!correct_format_vec(pars_init)) {
-    stop("pars_init must be a named list corresponding to the parameters being sampled")
-  }
-  if(!correct_format_vec(pars_min)) {
-    stop("pars_min must be a named list corresponding to the parameters being sampled")
-  }
-  
-  if(!correct_format_vec(pars_max)) {
-    stop("pars_max must be a named list corresponding to the parameters being sampled")
-  }
-  if(!correct_format_vec(pars_discrete)) {
-    stop("pars_discrete must be a named list corresponding to the parameters being sampled")
-  }
-  
-  if(!correct_format_mat(proposal_kernel)) {
+
+  if(!correct_format_mat(proposal_kernel, par_names)) {
     stop("proposal_kernel must be a matrix or vector with names corresponding to the parameters being sampled")
   }
   
-  if(!is.logical(output_proposals) || length(output_proposals) != 1) {
-    stop("output_proposals must be either TRUE or FALSE")
+  # Convert data frame into named lists
+  df_col_to_list <- function(df, column) {
+    list_out <- as.list(df[[column]])
+    names(list_out) <- as.character(df$names)
+    list_out
   }
-  
 
-    if (length(model_params$beta_y) > 1) {
-      stop("Set beta variation through generate_beta_func in sircovid_model, not model_params")
+  pars_init <- df_col_to_list(pars_to_sample, "init")
+  pars_min <- df_col_to_list(pars_to_sample, "min")
+  pars_max <- df_col_to_list(pars_to_sample, "max")
+  pars_discrete <- df_col_to_list(pars_to_sample, "discrete")
+
+  is.numeric.list <- function(x) all(vapply(X = x, is.numeric, logical(1)))
+  is.logical.list <- function(x) all(vapply(X = x, is.logical, logical(1)))
+  
+  if(!is.numeric.list(pars_init)) {
+    stop('pars_min entries must be numeric')
+  }
+  if(!is.numeric.list(pars_min)) {
+    stop('pars_min entries must be numeric')
+  }
+  if(!is.numeric.list(pars_max)) {
+    stop('pars_max entries must be numeric')
+  }
+
+  if(!is.logical.list(pars_discrete)) {
+    stop('pars_discrete entries must be logical')
+  }
+
+  if(any(pars_to_sample$init < pars_to_sample$min | pars_to_sample$init > pars_to_sample$max)) {
+    stop('initial parameters are outside of specified range')
+  }
+  if(pars_init['beta_start'] < 0) {
+    stop('beta_start must not be negative')
+  }
+  if('beta_end' %in% par_names) {
+    if(pars_init['beta_end'] < 0) {
+      stop('beta_end must not be negative')
     }
+  }
+  if(start_date_to_offset(data$date[1], pars_init[['start_date']]) < 0) {
+    stop('start date must not be before first date of supplied data')
+  }
 
   #
   # Generate MCMC parameters
   #
+  
   inputs <- list(
     data = data,
     n_mcmc = n_mcmc,
@@ -227,88 +253,49 @@ pmcmc <- function(data,
     n_particles = n_particles, 
     steps_per_day = steps_per_day)
   
-  # convert dates to be Date objects
-  data$date <- as.Date(data$date) 
-  
-  is.numeric.list <- function(x) all(vapply(X = x, is.numeric, logical(1)))
-  is.logical.list <- function(x) all(vapply(X = x, is.logical, logical(1)))
-  
-  if(!is.numeric.list(pars_min)) {
-    stop('pars_min entries must be numeric')
-  }
-  if(!is.numeric.list(pars_max)) {
-    stop('pars_max entries must be numeric')
-  }
-
-  if(!is.logical.list(pars_discrete)) {
-    stop('pars_discrete entries must be logical')
-  }
-  
   # needs to be a vector to pass to reflecting boundary function
   pars_min <- unlist(pars_min)  
   pars_max <- unlist(pars_max)
   pars_discrete <- unlist(pars_discrete)
-  
-  #
-  # create prior and likelihood functions given the inputs
-  #
-  if(is.null(log_prior)) {
-    # set improper, uninformative prior
-    log_prior <- function(pars) log(1e-10)
-  }
-  calc_lprior <- log_prior
-  
-  if(is.null(log_likelihood)) {
-    log_likelihood <- calc_loglikelihood
-  } else if (!('...' %in% names(formals(log_likelihood)))){
-    stop('log_likelihood function must be able to take unnamed arguments')
-  }
+  curr_pars <- unlist(pars_init)
+
+  # convert the current parameters into format easier for mcmc to deal with
+  curr_pars['start_date'] <- start_date_to_offset(data$date[1], pars_init$start_date)
   
   # create shorthand function to calc ll given main inputs
   # binds these input parameters
   calc_ll <- function(pars) {  
-    X <- log_likelihood(pars = pars, 
-                        data = data,
-                        sircovid_model = sircovid_model,
-                        model_params = model_params,
-                        steps_per_day = steps_per_day, 
-                        pars_obs = pars_obs, 
-                        n_particles = n_particles,
-                        forecast_days = 0,
-                        return = "ll"
+    X <- calc_loglikelihood(pars = pars, 
+                            data = data,
+                            sircovid_model = sircovid_model,
+                            model_params = model_params,
+                            steps_per_day = steps_per_day, 
+                            pars_obs = pars_obs, 
+                            n_particles = n_particles,
+                            forecast_days = 0,
+                            return = "ll"
     ) 
     X
   }
-  
+
+  # same thing for priors
+  calc_lprior <- function(pars) {
+    lprior <- 0
+    for (par in names(pars)) {
+      lprior <- lprior + pars_lprior[[par]](pars) 
+    }
+    lprior
+  }
+
   # create shorthand function to propose new pars given main inputs
   propose_jump <- function(pars) {
     propose_parameters(pars = pars, 
-                       proposal_kernel = proposal_kernel,
+                       proposal_kernel = proposal_kernel[names(pars), names(pars)],
                        pars_discrete = pars_discrete,
                        pars_min = pars_min,
                        pars_max = pars_max)
   }
-  
-  # convert the current parameters into format easier for mcmc to deal with
-  curr_pars <- unlist(pars_init)
-  curr_pars['start_date'] <- start_date_to_offset(data$date[1], pars_init$start_date) # convert to numeric
-  
-  if(any(curr_pars < pars_min | curr_pars > pars_max)) {
-    stop('initial parameters are outside of specified range')
-  }
-  if(curr_pars['beta_start'] < 0) {
-    stop('beta_start must not be negative')
-  }
-  if('beta_end' %in% pars_to_sample) {
-    if(curr_pars['beta_end'] < 0) {
-      stop('beta_end must not be negative')
-    }
-  }
-  if(curr_pars['start_date'] < 0) {
-    stop('start date must not be before first date of supplied data')
-  }
 
-  # Run the chains in parallel
   chains <- furrr::future_pmap(
       .l =  list(n_mcmc = rep(n_mcmc, n_chains)), 
       .f = run_mcmc_chain,
@@ -329,7 +316,7 @@ pmcmc <- function(data,
     chains_coda <- lapply(chains, function(x) {
         
         traces <- x$results
-        if('start_date' %in% pars_to_sample) {
+        if('start_date' %in% pars_to_sample$names) {
           traces$start_date <- start_date_to_offset(data$date[1], traces$start_date)
         }
         
@@ -385,23 +372,6 @@ run_mcmc_chain <- function(inputs,
 
   if(is.infinite(curr_lprior)) {
     stop('initial parameters are not compatible with supplied prior')
-  }
-  
-  
-  if(length(p_filter_est) != 2) {
-    stop('log_likelihood function must return a list containing elements log_likelihood and sample_state')
-  }
-  if(!setequal(names(p_filter_est), c('log_likelihood', 'sample_state'))) {
-    stop('log_likelihood function must return a list containing elements log_likelihood and sample_state')
-  }
-  if(length(p_filter_est$log_likelihood) > 1) {
-    stop('log_likelihood must be a single numeric representing the estimated log likelihood')
-  }
-  if(p_filter_est$log_likelihood > 0) {
-    stop('log_likelihood must be negative or zero')
-  }
-  if(any(p_filter_est$sample_state < 0)) {
-    stop('sample_state must be a vector of non-negative numbers')
   }
   
   # extract loglikelihood estimate and sample state
@@ -534,29 +504,32 @@ calc_loglikelihood <- function(pars, data, sircovid_model, model_params,
     stop("Unknown return type to calc_loglikelihood")
   }
   
-  # pars[['start_date']] argument is an integer reflecting the number of days between 
-  # the model start date and the first date in the data
-  if ('start_date' %in% names(pars)) {
-    start_date <- offset_to_start_date(data$date[1], pars[['start_date']])
-  } else {
-    start_date <- data$date[1]
-  }
-  if ('beta_end' %in% names(pars)) {
-    beta_end <- pars[['beta_end']]
-  } else {
-    beta_end <- NULL
-  }
-  if ('beta_start' %in% names(pars)) {
-    beta_start <- pars[['beta_start']]
-  } else {
-    beta_start <- NULL
-  }
-  if ('beta_pl' %in% names(pars)) {
-    beta_pl <- pars[['beta_pl']]
-  } else {
-    beta_pl <- NULL
-  }
+  # defaults if not being sampled
+  beta_start <- NULL
+  beta_end <- NULL
+  beta_pl <- NULL
+  start_date <- data$date[1]
   
+  # Update particle filter parameters from pars
+  for (par in names(pars)) {
+    if (par == "start_date") {
+      start_date <- offset_to_start_date(data$date[1], pars[[par]])
+    } else if (par == "beta_start") {
+      beta_start <- pars[[par]]
+    } else if (par == "beta_end") {
+      beta_end <- pars[[par]]
+    } else if (par == "beta_pl") {
+      beta_pl <- pars[[par]]
+    } else if (par %in% names(model_params)) {
+      model_params[[par]] <- pars[[par]]
+    } else if (par %in% names(pars_obs)) {
+      pars_obs[[par]] <- pars[[par]]
+    } else {
+      stop(paste0("Don't know how to update parameter: ", par))
+    }
+  }
+
+  # Beta needs a transform applied
   new_beta <- update_beta(sircovid_model, 
                           beta_start, 
                           beta_end, 
@@ -565,13 +538,6 @@ calc_loglikelihood <- function(pars, data, sircovid_model, model_params,
                           model_params$dt)
   model_params$beta_y <- new_beta$beta_y
   model_params$beta_t <- new_beta$beta_t
-  
-  fitted_gammas <- names(pars)[grep('gamma', names(pars))]
-  
-  for(par in fitted_gammas) {
-    model_params[[par]] <- pars[[par]]
-  }
-  
 
   pf_result <- run_particle_filter(data = data,
                                    sircovid_model = sircovid_model,
@@ -648,7 +614,7 @@ summary.pmcmc <- function(object, ...) {
   par_names <- names(object$inputs$pars$pars_init)
   
   ## convert start_date to numeric to calculate stats
-  data_start_date <- as.Date(object$inputs$data$date[1])
+  data_start_date <- sircovid_date(object$inputs$data$date[1])
   traces <- object$results[,par_names] 
   traces$start_date <- start_date_to_offset(data_start_date, traces$start_date)
   
@@ -674,7 +640,8 @@ summary.pmcmc <- function(object, ...) {
   
   sds <- round(apply(traces, 2, sd), 3)
   # convert start_date back into dates
-  summ$start_date <- as.Date(-summ$start_date, data_start_date)
+  summ$start_date <- offset_to_start_date(data_start_date, summ$start_date)
+  summ$start_date <- as.Date(summ$start_date, origin="2019-12-31")
   summ[c('2.5%', '97.5%', 'min', 'max'), 'start_date'] <- summ[c('97.5%', '2.5%', 'max', 'min'), 'start_date']